requirements.txt
Flask==3.0.0
gunicorn==21.2.0
pymongo[srv]==4.6.0
python-dotenv==1.0.0
dnspython==2.4.2